NetworkResource: Add `PreferIpv6` configuration option
Add an option to a job's network configuration to hint it to prefer IPv6
in a dual stack environment. In the current implementation, when a dual
stack CNI network initializes, only the IPv4 address is reported to
consul for service registration. When this flag is enabled, the IPv6
address takes precedence over an IPv4 address for service registration
when both are present.